PlantDoctor-website/requirements.txt
flask==2.0.1
SQLAlchemy==1.4
flask-sqlalchemy==2.5.1
flask-cors==3.0.10
python-dotenv==0.19.0
pyjwt==2.1.0
werkzeug==2.0.1